refactor(astra-sandbox): extract tool_sandbox module into standalone crate
Extract the tool_sandbox module (~3.4K LOC) from runtime into a new
astra-sandbox crate. This continues the modularization effort.

Changes:
- Create astra-sandbox crate with 9 modules:
  - bash_ast: Bash AST parsing for command safety analysis
  - command: Command sandboxing (env filtering, limits)
  - git_safety: Git command validation
  - path: Path boundary validation
  - policy: SandboxPolicy and SandboxMode
  - process_isolation: Isolated subprocess execution
  - shell_hardening: Shell environment hardening
  - tier: ToolTier classification for execution routing

- Define local TrustTier enum to avoid circular dependency with
  astra-skills (not yet merged)

- Runtime now depends on astra-sandbox and re-exports its types
- Added From impl for TrustTier conversion at runtime boundary
- Removed duplicate files from runtime (~3.4K LOC reduction)

Test: 179 astra-sandbox tests + 293 runtime tests pass
