+++ rust/Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "crates/astra-edge",
     "crates/astra-pipeline",
     "crates/astra-turn-types",
+    "crates/astra-sandbox",
 ]
 resolver = "2"
 
@@ -20,6 +21,7 @@ edition = "2024"
 [workspace.dependencies]
 astra-core = { path = "crates/core" }
 astra-pipeline = { path = "crates/astra-pipeline" }
+astra-sandbox = { path = "crates/astra-sandbox" }
 astra-turn-types = { path = "crates/astra-turn-types" }
 astra-tools = { path = "crates/astra-tools" }
 astra-services = { path = "crates/services" }
