+++ rust/crates/astra-sandbox/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "astra-sandbox"
+version = "0.1.0"
+edition = "2024"
+description = "Security boundary enforcement for Astra agent tool execution"
+
+[dependencies]
+# Core serialization
+serde = { workspace = true }
+serde_json = { workspace = true }
+regex = { workspace = true }
+
+# Async runtime
+tokio = { workspace = true, features = ["process", "io-util", "time"] }
+
+# Logging
+tracing = { workspace = true }
+
+# AST parsing for bash safety
+tree-sitter = "0.24"
+tree-sitter-bash = "0.23"
+
+# UUID for isolation namespaces
+uuid = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
+tempfile = "3"
