+++ rust/crates/runtime/Cargo.toml
@@ -32,6 +32,7 @@ dirs = "6.0.0"
 fastrand = "2"
 astra-core.workspace = true
 astra-pipeline.workspace = true
+astra-sandbox.workspace = true
 astra-turn-types.workspace = true
 astra-services.workspace = true
 astra-tools.workspace = true
@@ -60,8 +61,6 @@ tracing-subscriber.workspace = true
 http-body-util.workspace = true
 uuid.workspace = true
 thiserror.workspace = true
-tree-sitter = "0.24"
-tree-sitter-bash = "0.23"
 notify = { version = "7", default-features = false, features = ["macos_kqueue"] }
 dashmap = "6.1.0"
 toml = { version = "1.1.2", features = ["preserve_order"] }
